[upstream_utils] Upgrade to fmt 9.0.0
fmt removed fmt::make_args_checked since it's no longer needed for
constexpr format string checks.

fmt deprecated implicit conversions from enums to integers in format
arguments, so I added explicit static casts.

+++ ntcore/src/main/native/cpp/Dispatcher.cpp
@@ -632,7 +632,7 @@ bool DispatcherBase::ClientHandshake(
       DEBUG0(
           "client: received message ({}) other than entry assignment during "
           "initial handshake",
-          msg->type());
+          static_cast<int>(msg->type()));
       return false;
     }
     incoming.emplace_back(std::move(msg));
@@ -735,7 +735,7 @@ bool DispatcherBase::ServerHandshake(
         DEBUG0(
             "server: received message ({}) other than entry assignment during "
             "initial handshake",
-            msg->type());
+            static_cast<int>(msg->type()));
         return false;
       }
       incoming.push_back(msg);
